added node_utils file
